fix: 字典管理修复，删除子项需要等到接口确认成功后，才能进行页面数据的删除

+++ packages/provider-app-hub/DictionaryManager/components/DictionaryList.tsx
@@ -96,6 +96,7 @@ const getChildOfDictionaryColumns = ({
               className="link-btn"
               onClick={(e) => {
                 delChildOfDictionaryServices({ dictionaryId, pid: id }).then((canIDelete) => {
+                  if (!canIDelete) return;
                   /** 删除成功后需要刷新子项列表 */
                   handleDeleteChild(record);
                 });
